src/lambda_codebase/initial_commit/bootstrap_repository/adf-build/shared/requirements.txt
# Install libs here that you might want in AWS CodeBuild
pytest==3.0.7
mock==2.0.0
boto3~=1.10, >=1.10.47
pyyaml>=5.1
schema==0.7.1
jsii==1.1.0
aws_cdk.aws_codepipeline==1.50
aws_cdk.aws_lambda_event_sources==1.50
aws-cdk.aws-codepipeline-actions==1.50
aws_cdk.aws_codebuild==1.50
aws_cdk.aws_codecommit==1.50
aws-cdk.aws-s3==1.50
aws-cdk-assets==1.50
aws-cdk.aws-iam==1.50
aws-cdk.aws-kms==1.50
aws-cdk.aws-lambda==1.50
aws-cdk.aws-sns==1.50
aws-cdk.aws_secretsmanager==1.50
aws-cdk.aws_ssm==1.50
aws-cdk.aws_events==1.50
aws_cdk.aws_cloudformation==1.50
aws_cdk.aws_certificatemanager==1.50
aws_cdk.aws_sqs==1.50
aws_cdk.aws_ec2==1.50
aws_cdk.aws_sns==1.50
aws_cdk.aws_elasticloadbalancing==1.50
aws_cdk.aws_s3_notifications==1.50
aws_cdk.aws_codedeploy==1.50
aws_cdk.aws_ecr==1.50
aws_cdk.aws_stepfunctions==1.50
aws_cdk.aws_servicediscovery==1.50
aws_cdk.aws_autoscaling_hooktargets==1.50
aws_cdk.aws_cloudfront==1.50
aws_cdk.aws_route53_targets==1.50
aws_cdk.aws_route53==1.50
aws_cdk.aws_ecs==1.50
aws_cdk.aws_events_targets==1.50
aws_cdk.aws_dynamodb==1.50
aws_cdk.aws_applicationautoscaling==1.50
aws_cdk.aws_autoscaling==1.50
aws_cdk.aws_autoscaling_common==1.50
aws_cdk.aws_s3_assets==1.50
aws_cdk.aws_ecr_assets==1.50
aws_cdk.aws_elasticloadbalancingv2==1.50
aws_cdk.aws_apigateway==1.50
aws_cdk.aws_logs==1.50
aws_cdk.aws_ecr==1.50
aws_cdk.aws_cloudwatch==1.50
aws_cdk.aws_kinesis==1.50
aws_cdk.aws_s3_assets==1.50
aws_cdk.aws_sns_subscriptions==1.50
aws-cdk.aws-sam==1.50
aws-cdk.cx-api==1.50
aws-cdk.region-info==1.50
aws-cdk.core==1.50
aws-cdk.pipelines==1.50
